Refactor database code to allow easier reuse with different backends
