+++ nano/secure/blockstore.cpp
@@ -365,30 +365,6 @@ void nano::representative_visitor::state_block (nano::state_block const & block_
 	result = block_a.hash ();
 }
 
-nano::uint128_t nano::block_store::block_balance_calculated (std::shared_ptr<nano::block> block_a, nano::block_sideband const & sideband_a) const
-{
-	nano::uint128_t result;
-	switch (block_a->type ())
-	{
-		case nano::block_type::open:
-		case nano::block_type::receive:
-		case nano::block_type::change:
-			result = sideband_a.balance.number ();
-			break;
-		case nano::block_type::send:
-			result = boost::polymorphic_downcast<nano::send_block *> (block_a.get ())->hashables.balance.number ();
-			break;
-		case nano::block_type::state:
-			result = boost::polymorphic_downcast<nano::state_block *> (block_a.get ())->hashables.balance.number ();
-			break;
-		case nano::block_type::invalid:
-		case nano::block_type::not_a_block:
-			release_assert (false);
-			break;
-	}
-	return result;
-}
-
 nano::read_transaction::read_transaction (std::unique_ptr<nano::read_transaction_impl> read_transaction_impl) :
 impl (std::move (read_transaction_impl))
 {
